+++ src/app/api/position/[id]/route.ts
@@ -1,4 +1,4 @@
-import { positionController } from '@/lib/controllers/position.controller';
+import { getPosition, deletePosition, updatePosition } from '@/lib/services/position.service';
 import { sargeApiError, sargeApiResponse } from '@/lib/responses';
 import { PositionNotFoundError } from '@/lib/schemas/position.schema';
 import { type NextRequest } from 'next/server';
@@ -6,7 +6,7 @@ import { type NextRequest } from 'next/server';
 export async function GET(_request: NextRequest, { params }: { params: Promise<{ id: string }> }) {
     try {
         const userId = (await params).id;
-        const user = await positionController.get(userId);
+        const user = await getPosition(userId);
         return sargeApiResponse(user, 200);
     } catch (error) {
         if (error instanceof PositionNotFoundError) {
@@ -24,7 +24,7 @@ export async function DELETE(
 ) {
     try {
         const positionId = (await params).id;
-        const deletedPosition = await positionController.delete(positionId);
+        const deletedPosition = await deletePosition(positionId);
         return sargeApiResponse(deletedPosition, 200);
     } catch (error) {
         if (error instanceof PositionNotFoundError) {
@@ -41,7 +41,7 @@ export async function PUT(request: NextRequest, { params }: { params: Promise<{ 
         const positionId = (await params).id;
         const positionData = await request.json();
 
-        const updatedPosition = await positionController.update(positionId, positionData);
+        const updatedPosition = await updatePosition(positionId, positionData);
         return sargeApiResponse(updatedPosition, 200);
     } catch (error) {
         if (error instanceof PositionNotFoundError) {
