+++ src/app/api/taskTemplates/[id]/route.ts
@@ -1,12 +1,16 @@
 import { sargeApiError, sargeApiResponse } from '@/lib/responses';
 import { InvalidInputError } from '@/lib/schemas/errors';
-import taskTemplateController from '@/lib/controllers/taskTemplate.controller';
+import {
+    getTaskTemplate,
+    deleteTaskTemplate,
+    updateTaskTemplate,
+} from '@/lib/services/task-template.service';
 import { TaskTemplateNotFoundError } from '@/lib/schemas/taskTemplate.schema';
 import { type NextRequest } from 'next/server';
 
 export async function GET(_req: NextRequest, { params }: { params: Promise<{ id: string }> }) {
     try {
-        const foundTaskTemplate = await taskTemplateController.get(await params);
+        const foundTaskTemplate = await getTaskTemplate(await params);
         return sargeApiResponse(foundTaskTemplate, 200);
     } catch (error) {
         if (error instanceof InvalidInputError) {
@@ -24,7 +28,7 @@ export async function GET(_req: NextRequest, { params }: { params: Promise<{ id:
 
 export async function DELETE(_req: NextRequest, { params }: { params: Promise<{ id: string }> }) {
     try {
-        const deletedTaskTemplate = await taskTemplateController.delete(await params);
+        const deletedTaskTemplate = await deleteTaskTemplate(await params);
         return sargeApiResponse(deletedTaskTemplate, 200);
     } catch (error) {
         if (error instanceof InvalidInputError) {
@@ -44,7 +48,7 @@ export async function PUT(request: NextRequest, { params }: { params: Promise<{ 
     try {
         const body = await request.json();
         const idWithBody = { id: (await params).id, ...body };
-        const taskTemplate = await taskTemplateController.update(idWithBody);
+        const taskTemplate = await updateTaskTemplate(idWithBody);
         return sargeApiResponse(taskTemplate, 200);
     } catch (error) {
         if (error instanceof InvalidInputError) {
