+++ src/app/api/users/route.ts
@@ -1,4 +1,4 @@
-import { userController } from '@/lib/controllers/user.controller';
+import { createUser } from '@/lib/services/user.service';
 import { type NextRequest } from 'next/server';
 import { sargeApiError, sargeApiResponse } from '@/lib/responses';
 import { InvalidInputError } from '@/lib/schemas/errors';
@@ -6,7 +6,7 @@ import { InvalidInputError } from '@/lib/schemas/errors';
 export async function POST(request: NextRequest) {
     try {
         const body = await request.json();
-        const user = await userController.create(body);
+        const user = await createUser(body);
         return sargeApiResponse(user, 200);
     } catch (error) {
         if (error instanceof InvalidInputError) {
